% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottomgraph.R
\name{bottomgraph}
\alias{bottomgraph}
\title{bottomgraph}
\usage{
bottomgraph(
  merged_table,
  conf,
  group1name,
  group2name,
  timeunit,
  timesig,
  data_long
)
}
\arguments{
\item{merged_table}{table from merged_results}

\item{conf}{confidence level}

\item{group1name}{string for name of group 1}

\item{group2name}{string for name of group 2}

\item{timeunit}{unit of time used as a string}

\item{timesig}{point of no significance from holm_test}

\item{data_long}{raw long-format, longitudinal data}
}
\value{
graph of pain over time by group
}
\description{
Creates plot showing rolling difference in cumulative AUC values over time
with first time of no significance highlighted
}
