% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_table.R
\name{cumulative_table}
\alias{cumulative_table}
\title{cumulative_table}
\usage{
cumulative_table(iterative_data)
}
\arguments{
\item{iterative_data}{matrix of individual, non-cumulative AUC data}
}
\value{
table with time, cumulative average, standard deviation,
and sample size
}
\description{
Intakes individual AUC data over time and calculates the cumulative average,
standard deviation, and sample size
}
