% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holm_test.R
\name{holm_test}
\alias{holm_test}
\title{holm_test}
\usage{
holm_test(x, conf)
}
\arguments{
\item{x}{table resulting from merged_results}

\item{conf}{confidence level}
}
\value{
time value of no statistical significance
}
\description{
Runs the Holm-Bonferoni method and extracts the first time of no statistical
significance
}
