% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_auc.R
\name{iterative_auc}
\alias{iterative_auc}
\title{iterative_auc}
\usage{
iterative_auc(data_long)
}
\arguments{
\item{data_long}{long format longitudinal data}
}
\value{
matrix of individual AUC per time interval
}
\description{
Imtakes raw long-format longitudinal data and calculates iterative AUC values
}
