% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_table.R
\name{iterative_table}
\alias{iterative_table}
\title{iterative_table}
\usage{
iterative_table(iterative_data)
}
\arguments{
\item{iterative_data}{matrix of individual AUC values}
}
\value{
table with time, average, standard deviation, and sample size by time interval
}
\description{
Intakes individual AUC data over time and calculates the average,
standard deviation, and sample size by time interval.
}
