% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_results.R
\name{merged_results}
\alias{merged_results}
\title{merged_results}
\usage{
merged_results(group1, group2)
}
\arguments{
\item{group1}{aggregated data from group1}

\item{group2}{aggregated data from group2}
}
\value{
merged table with extra metrics for both groups
}
\description{
Aggregates data from 2 individual tables into one and calculates the
mean difference, pooled variance, t-test statistic, and p-value
}
