% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painplot.R
\name{painplot}
\alias{painplot}
\title{painplot}
\usage{
painplot(
  raw_data,
  conf = 0.95,
  group1name,
  group2name,
  timeunit = "hours",
  completecase = FALSE
)
}
\arguments{
\item{raw_data}{raw long-format, longitudinal data}

\item{conf}{confidence level}

\item{group1name}{string for name of group 1}

\item{group2name}{string for name of group 2}

\item{timeunit}{unit of time used as a string}

\item{completecase}{option to exclude subjects without complete data}
}
\value{
2 visuals
}
\description{
Takes in long form longitudinal data and extra information to produce
2 visuals.
}
