% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topgraph.R
\name{topgraph}
\alias{topgraph}
\title{topgraph}
\usage{
topgraph(x, conf)
}
\arguments{
\item{x}{raw, long-format longitudinal data}

\item{conf}{confidence level}
}
\value{
graph of pain over time by group
}
\description{
Creates plot which visualizes pain scores longitudinally over time by group
}
